requirements.txt
# Forex Data Pipeline - Python Dependencies

# ========================================
# dbt and BigQuery
# ========================================
dbt-bigquery>=1.7.0

# ========================================
# Data Extraction and Processing
# ========================================
twelvedata>=1.2.0           # Twelve Data API client
pandas>=2.0.0               # Data manipulation and analysis
pyarrow>=14.0.0             # Parquet file support

# ========================================
# Google Cloud Platform
# ========================================
google-cloud-storage>=2.10.0   # GCS client library
google-cloud-bigquery>=3.11.0  # BigQuery client library

# ========================================
# Configuration
# ========================================
python-dotenv>=1.0.0        # Environment variable management
